+++ src/modules/sidebarProvider.ts
@@ -222,7 +222,7 @@ window.isDarkMode = ${isDarkMode};
 window.hiddenStatuses = ${JSON.stringify(Settings.sidebar.hiddenStatuses)};
 window.partyUri = '${Settings.sidebar.showAcGif ? getUri('res/party.gif') : ''}';
 window.language = '${env.language}';
-window.showTips = '${Settings.sidebar.showTips}';
+window.showTips = ${Settings.sidebar.showTips};
 window.fontFamily = \`${Settings.sidebar.fontFamily}\`;
 </script>
 </body>