+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
-## [Unreleased]
+## 0.4.3
 
 - **fix**: disable show tips
 - **fix**: disable auto focus
 - **feat**: add disable/enable test case feature
 - **chore**: upgrade packages
+- **fix**: update debounce import path in extensionManager
 
 ## 0.4.2